Implement Final Production Hardening, LHDN Connection Testing, and Hardware HAL Enhancements.
This update transitions the system to a 100% production-ready state, finalizing the Malaysian LHDN MyInvois integration and significantly improving hardware reliability and in-app merchant support documentation.

Key technical changes include:

- **LHDN e-Invoicing Hardening**: Refactored `InvoisMapper` to use `BigDecimal` precision for all financial fields, eliminating floating-point errors. Added a "Test LHDN Connection" feature in Settings and implemented comprehensive unit tests for document mapping and tax grouping logic.
- **Hardware Abstraction Layer (HAL) Refactor**: Introduced `PrinterFactory` to centralize printer instantiation. Expanded the ESC/POS engine to support bit-image printing (`GS v 0`) and dynamic character widths (32/42/48) based on paper size (58mm/80mm).
- **Payment Terminal Integration**: Added `GHL_JSON` terminal support, implementing a standard JSON-over-TCP protocol used by Malaysian payment acquirers for integrated card and QR transactions.
- **Printer Driver Reliability**:
    - **USB**: Implemented chunked transfers to prevent buffer overflow on low-cost thermal printers.
    - **Bluetooth**: Added retry logic and connection timeouts for more resilient pairing.
    - **Network**: Enhanced socket state tracking to prevent "broken pipe" errors during print jobs.
- **In-App Documentation System**: Added a multi-tab `HelpScreen` and `AboutScreen`. These modules pull from localized Markdown assets (`getting_started.md`, `lhdn_compliance.md`) to provide merchants with offline access to user manuals and changelogs.
- **UI & Navigation Consolidation**: Centralized sidebar navigation within `MainScreen` to resolve layout inconsistencies. Moved shared UI components like `NavButton` to a common package and updated the styling to a high-contrast "Slate" theme.
- **Financial & Analytics Precision**: Standardized `BigDecimal` usage across the `DiscountDialog`, `AnalyticsViewModel`, and P2P sync engine. Chart data points now include dedicated `displayValue` fields to ensure consistency between visual trends and numerical reports.
- **Deployment Readiness**: Added `DEPLOYMENT.md` and `INTEGRATIONS.md` to document the release process, device-bound licensing logic, and P2P synchronization protocols for technical deployments.

+++ app/src/main/java/com/extrotarget/extroposv2/core/util/lhdn/LhdnInvoicingUtils.kt
@@ -2,6 +2,7 @@ package com.extrotarget.extroposv2.core.util.lhdn
 
 import com.extrotarget.extroposv2.core.data.model.Sale
 import com.extrotarget.extroposv2.core.data.model.SaleItem
+import java.math.BigDecimal
 import java.security.MessageDigest
 import java.text.SimpleDateFormat
 import java.util.*
@@ -21,8 +22,8 @@ object LhdnInvoicingUtils {
     /**
      * For B2C Consolidated e-Invoices, LHDN requires grouping by tax rate.
      */
-    fun groupItemsByTax(items: List<SaleItem>): Map<Double, List<SaleItem>> {
-        return items.groupBy { it.taxRate.toDouble() }
+    fun groupItemsByTax(items: List<SaleItem>): Map<BigDecimal, List<SaleItem>> {
+        return items.groupBy { it.taxRate }
     }
 
     /**
